Framework for Display, Movement and Sensors
This Firmware is the basic Framework for the rover, the firmware holds all the sensor and output devices required for movement. The Rover takes Temp/Humidity , Light Intensity and scans for close objects displaying the results to the 4Digit 7Segment Display unit, When the IR Receiver detects an IR Signal, it processes the information, this information can be used to control the rover and is useful for changing "MODES" or state at which the Rover is at.